Change int types to double to reflect actual values returned by Shopify

+++ ShopifySharp/Services/Graph/GraphRequestCostThrottleStatusExtension.cs
@@ -2,9 +2,9 @@ namespace ShopifySharp;
 
 public class GraphRequestCostThrottleStatusExtension
 {
-    public int RestoreRate { get; set; }
+    public double RestoreRate { get; set; }
 
-    public int CurrentlyAvailable { get; set; }
+    public double CurrentlyAvailable { get; set; }
 
-    public int MaximumAvailable { get; set; }
+    public double MaximumAvailable { get; set; }
 }
